fixed render bug for when book is edited
